refactor: convertClovaEventStream2TokenStream 작성
